changes brew ST3 cask
The old versioned cask has been deprecated, and the cask sublime-text
now points to ST3 instead of ST2.

Also cleans up and adds notes to explain some of the previous changes
to this subscript.

Adds a step in the README to install Chrome, since that is no longer
auto-installed.

Builds the scripts.